Solutions/Feedly/Data Connectors/requirements.txt
azure-functions
azure-storage-file-share==12.3.0
feedly-client==0.26
marshmallow-dataclass==8.5.14
pandas==2.0.3